el controlador y l vista de añadir fueron adaptados #25

+++ src/main/java/com/storecode/repositories/IUserRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.storecode.models.User;
 
-public interface IUserRepository extends JpaRepository<User, Integer>{
+public interface IUserRepository extends JpaRepository<User, Long>{
 
 }
